move jsonPrettify to insomnia src
