--- packages/insomnia-prettify/.eslintignore
@@ -1,1 +1,0 @@
-dist
