--- packages/insomnia-prettify/README.md
@@ -1,5 +1,0 @@
-# Insomnia Prettify
-
-[![Npm Version](https://img.shields.io/npm/v/insomnia-prettify.svg)](https://www.npmjs.com/package/insomnia-prettify)
-
-Prettification utilities for various formats.
