--- packages/insomnia-prettify/jest.config.js
@@ -1,9 +1,0 @@
-/** @type { import('@jest/types').Config.InitialOptions } */
-module.exports = {
-  preset: '../../jest-preset.js',
-  globals: {
-    'ts-jest': {
-      isolatedModules: true,
-    },
-  },
-};
