--- packages/insomnia-prettify/package.json
@@ -1,36 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-prettify",
-  "version": "3.6.1-beta.3",
-  "author": "Kong <office@konghq.com>",
-  "description": "Prettification utilities for Insomnia",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "packages/insomnia-prettify"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "package.json",
-    "README.md"
-  ],
-  "devDependencies": {
-    "@jest/globals": "^28.1.0",
-    "jest": "^28.1.0"
-  },
-  "scripts": {
-    "bootstrap": "npm run build",
-    "lint": "eslint . --ext .js,.ts,.tsx --cache",
-    "lint:fix": "npm run lint -- --fix",
-    "clean": "tsc --build tsconfig.build.json --clean",
-    "postclean": "rimraf dist",
-    "build": "tsc --build tsconfig.build.json",
-    "test": "jest"
-  }
-}
