--- packages/insomnia-prettify/src/index.ts
@@ -1,1 +1,0 @@
-export { prettify as json } from './json';
