+++ packages/insomnia/package.json
@@ -84,7 +84,6 @@
     "insomnia-plugin-request": "^3.6.1-beta.3",
     "insomnia-plugin-response": "^3.6.1-beta.3",
     "insomnia-plugin-uuid": "^3.6.1-beta.3",
-    "insomnia-prettify": "^3.6.1-beta.3",
     "insomnia-testing": "^3.6.1-beta.3",
     "insomnia-url": "^3.6.1-beta.3",
     "insomnia-xpath": "^3.6.1-beta.3",
