+++ packages/insomnia/src/ui/components/editors/body/graph-ql-editor.tsx
@@ -8,7 +8,6 @@ import { DefinitionNode, DocumentNode, GraphQLNonNull, GraphQLSchema, Kind, NonN
 import { buildClientSchema, getIntrospectionQuery } from 'graphql/utilities';
 import { Maybe } from 'graphql-language-service';
 import { jarFromCookies } from 'insomnia-cookies';
-import { json as jsonPrettify } from 'insomnia-prettify';
 import { buildQueryStringFromParams, joinUrlAndQueryString, setDefaultProtocol } from 'insomnia-url';
 import prettier from 'prettier';
 import React, { FC, useEffect, useRef, useState } from 'react';
@@ -22,6 +21,7 @@ import * as models from '../../../../models';
 import type { Request } from '../../../../models/request';
 import type { Settings } from '../../../../models/settings';
 import { axiosRequest } from '../../../../network/axios-request';
+import { jsonPrettify } from '../../../../utils/prettify/json';
 import { Dropdown } from '../../base/dropdown/dropdown';
 import { DropdownButton } from '../../base/dropdown/dropdown-button';
 import { DropdownDivider } from '../../base/dropdown/dropdown-divider';
