+++ packages/insomnia/src/utils/prettify/fixtures/extra-whitespace-input.json
