+++ packages/insomnia/src/utils/prettify/fixtures/extra-whitespace-output.json
