+++ packages/insomnia/src/utils/prettify/fixtures/minimal-whitespace-input.json
