+++ packages/insomnia/src/utils/prettify/fixtures/minimal-whitespace-output.json
