+++ packages/insomnia/src/utils/prettify/fixtures/trailing-comma-input.json
