+++ packages/insomnia/src/utils/prettify/fixtures/trailing-comma-output.json
